polymaker_botV2.28/requirements.txt
# Polymaker 量化机器人依赖
# Python 3.11+

# ============================================
# 核心依赖
# ============================================

# HTTP 请求
requests>=2.28.0

# 时区处理
pytz>=2023.3

# 环境变量
python-dotenv>=1.0.0

# ============================================
# 以太坊相关
# ============================================

# Web3
web3>=6.0.0

# 以太坊账户
eth-account>=0.10.0

# ============================================
# Polymarket 官方 SDK
# ============================================

# CLOB 客户端
py-clob-client>=0.1.0

# Builder 签名 SDK
py-builder-signing-sdk>=0.0.2

# Builder Relayer 客户端（Gasless 交易）
py-builder-relayer-client>=0.0.1

# ============================================
# 数据处理
# ============================================

# Pydantic 数据验证
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 异步 SQLite
aiosqlite>=0.19.0

# Pandas 数据处理
pandas>=2.0.0

# ============================================
# Telegram Bot
# ============================================

# aiogram v3 (异步 Telegram Bot)
aiogram>=3.11.0

# ============================================
# 可选依赖
# ============================================

# 结构化日志（可选）
# structlog>=23.0.0
